mv browser client to client folder
