+++ node/src/lib.rs
@@ -1,8 +1,6 @@
 #![doc = include_str!("../README.md")]
 
 pub mod backend;
-#[cfg(feature = "browser")]
-pub mod browser;
 pub mod client;
 pub mod consts;
 pub mod error;
